+++ tests/arrays/build.test.ts
@@ -12,7 +12,7 @@ const x: int[]
   const parsed = parse(input);
   const result = build(parsed.root);
   const expected = `
-int x[];
+long x[];
 `;
   expect(parsed.errors).toEqual([]);
   expect(trim_test_build(result.code)).toEqual(trim_test_build(expected));
@@ -25,7 +25,7 @@ var x = [1, 2, 3]
   const parsed = parse(input);
   const result = build(parsed.root);
   const expected = `
-int x[3] = {1, 2, 3};
+long x[3] = {1, 2, 3};
 `;
   expect(parsed.errors).toEqual([]);
   expect(trim_test_build(result.code)).toEqual(trim_test_build(expected));
